make bufferId 16-bit
allows for more command buffers to be used

this may help when using buffers for logic and programming - being limited to 256 buffers could have proven too restrictive.  moving to a 16-bit identifier gives room to breath and space out bufferIds

also add support for clearing all buffers by sending the “clear buffer” command with a buffer ID of 65535

+++ video/vdu_stream_processor.h
@@ -73,10 +73,10 @@ class VDUStreamProcessor {
 		void receiveBitmap(uint8_t cmd, uint16_t width, uint16_t height);
 
 		void vdu_sys_buffered();
-		void bufferWrite(uint8_t bufferId);
-		void bufferCall(uint8_t bufferId);
-		void bufferClear(uint8_t bufferId);
-		void bufferAdjust(uint8_t bufferId);
+		void bufferWrite(uint16_t bufferId);
+		void bufferCall(uint16_t bufferId);
+		void bufferClear(uint16_t bufferId);
+		void bufferAdjust(uint16_t bufferId);
 };
 
 // Read an unsigned byte from the serial port, with a timeout
